chore(ci): migrate AI eval summary to CI report section
Rewrite post-eval-summary.js (github-script createComment per push) as
post-eval-section.mjs posting into the shared CI report comment.
Regression/improvement categorization and ordering are preserved; the
section warns when any experiment regressed. Expands the ci-ai
sparse-checkout to include the CI report engine.

Generated-By: PostHog Code
Task-Id: 47cda15c-e2d1-43f3-acca-31b57bbb90be
