+++ .github/workflows/ci-ai.yml
@@ -157,7 +157,9 @@ jobs:
         steps:
             - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
               with:
-                  sparse-checkout: .github/scripts
+                  sparse-checkout: |
+                      .github/scripts
+                      frontend/bin/ci-report
 
             - name: Download all eval results
               uses: actions/download-artifact@37930b1c2abaa49bbe596cd826c3c89aef350131 # v7.0.0
@@ -176,15 +178,17 @@ jobs:
                       wc -l eval_results.jsonl
                   fi
 
-            - name: Post eval summary to PR
+            - name: Post eval summary to CI report comment
               # always() because we want to post even if some eval groups errored
               if: always() && github.event_name == 'pull_request'
-              uses: actions/github-script@ed597411d8f924073f98dfc5c65a23a2325f34cd # v8.0.0
-              with:
-                  script: |
-                      const fs = require("fs")
-                      const script = require('.github/scripts/post-eval-summary.js')
-                      script({ github, context, fs })
+              env:
+                  GITHUB_TOKEN: ${{ github.token }}
+              run: |
+                  if [ -f .github/scripts/post-eval-section.mjs ]; then
+                      node .github/scripts/post-eval-section.mjs
+                  else
+                      echo "Skipping eval section — script not present on this checkout"
+                  fi
 
             - name: Upload merged test results
               uses: actions/upload-artifact@b7c566a772e6b6bfb58ed0dc250532a479d7789f # v6.0.0
